requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
# LangChain 1.x moved classic chains; pin to 0.3.x for stable imports used in this project.
langchain>=0.3.20,<1.0.0
langchain-community>=0.3.20,<1.0.0
langchain-openai>=0.2.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
openai>=1.0.0
fpdf2>=2.7.0
